requirements.txt
fastapi~=0.115.5
python-dotenv~=1.0.1
PyPDF2~=3.0.1